requirements.txt
Django>=4.0
djangorestframework>=3.12,<4.0
django-filter>=2.4,<3.0